Make masterLayout
